style(app): rename all css file scss
