+++ packages/app/app/components/AlbumCover/AlbumInfo/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from './styles.css';
+import styles from './styles.scss';
 
 const AlbumInfo = ({ artist, title }) => (
   <div className={styles.album_info_container}>
