+++ packages/app/app/components/AlbumCover/AlbumInfo/styles.scss
