+++ packages/app/app/components/AlbumCover/AlbumOverlay/styles.scss
