+++ packages/app/app/components/AlbumCover/styles.scss
