+++ packages/app/app/components/Dashboard/BestNewMusicTab/index.js
@@ -15,7 +15,7 @@ class BestNewMusicTab extends React.Component {
     };
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (this.state.activeItem === null) {
       const firstAlbum = _.head(
         _.get(nextProps, 'dashboardData.bestNewAlbums')
@@ -35,7 +35,7 @@ class BestNewMusicTab extends React.Component {
 
   setActiveItem(activeItem) {
     this.setState({ activeItem });
-    document.getElementsByClassName("best_new_music_content")[0].scrollTo(0,0);
+    document.getElementsByClassName('best_new_music_content')[0].scrollTo(0, 0);
   }
 
   render () {
