+++ packages/app/app/components/WindowControls/WindowButton/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import FontAwesome from 'react-fontawesome';
 
-import styles from './styles.css';
+import styles from './styles.scss';
 
 const WindowButton = ({ color, icon, onClick }) => (
   <div
