+++ packages/app/app/components/WindowControls/WindowButton/styles.scss
