+++ packages/app/app/components/WindowControls/styles.scss
